Initial Index.html commit
This is a test file to see if I can drop in the Horizontal Navigation Menu from my website latechdirect.com using some Jquery, into this index page and thus into the main page on my GitHub Page. It deals with cross-site scripting so this may not work due to browser security rules. It will be most likely removed from the final Joomla Module that I am creating, so this is just a test. We shall see [@latechdirect]